+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat: Replace NSString for eventId with SentryId #668
 - fix: Umbrella header #671
 - feat: Bump minimum iOS version to 9.0 #669
 - feat: Remove deprecated SDK inits #673